+++ src/lib.rs
@@ -1,22 +1,31 @@
 #![deny(unsafe_code)]
-//! Rust enums are great for types where all variations are known beforehand. But in
-//! the case where you want to implement a container of user-defined types, an
-//! open-ended type like a **trait object** is needed. In some cases, it is useful to
-//! cast the trait object back into its original concrete type to access additional
-//! functionality and performant inlined implementations.
+//! Rust enums are great for types where all variations are known beforehand. But a
+//! container of user-defined types requires an open-ended type like a **trait
+//! object**. Some applications may want to cast these trait objects back to the
+//! original concrete types to access additional functionality and performant
+//! inlined implementations.
 //!
-//! `downcast-rs` adds downcasting support to trait objects using only safe Rust. It
-//! supports **type parameters**, **associated types**, and **constraints**.
+//! `downcast-rs` adds this downcasting support to trait objects using only safe
+//! Rust. It supports **type parameters**, **associated types**, and **constraints**.
 //!
-//! To make a trait downcastable, make it extend the `downcast::Downcast` trait and
-//! invoke `impl_downcast!` on it as follows:
+//! To make a trait downcastable, make it extend either `downcast::Downcast` or
+//! `downcast::DowncastSync` and invoke `impl_downcast!` on it as in the examples
+//! below.
+//!
+//! Since 1.1.0, the minimum supported Rust version is 1.29 to support downcasting
+//! `Rc` and `Arc`.
 //!
 //! ```
 //! # #[macro_use]
 //! # extern crate downcast_rs;
-//! # use downcast_rs::Downcast;
+//! # use downcast_rs::{Downcast, DowncastSync};
 //! trait Trait: Downcast {}
 //! impl_downcast!(Trait);
+//!
+//! // Also supports downcasting `Arc`-ed trait objects by extending `DowncastSync`
+//! // and starting `impl_downcast!` with `sync`.
+//! trait TraitSync: DowncastSync {}
+//! impl_downcast!(sync TraitSync);
 //!
 //! // With type parameters.
 //! trait TraitGeneric1<T>: Downcast {}
@@ -44,15 +53,17 @@
 //! # Example without generics
 //!
 //! ```
+//! # use std::rc::Rc;
+//! # use std::sync::Arc;
 //! // Import macro via `macro_use` pre-1.30.
 //! #[macro_use]
 //! extern crate downcast_rs;
-//! use downcast_rs::Downcast;
+//! use downcast_rs::DowncastSync;
 //!
-//! // To create a trait with downcasting methods, extend `Downcast` and run
-//! // `impl_downcast!()` on the trait.
-//! trait Base: Downcast {}
-//! impl_downcast!(Base);
+//! // To create a trait with downcasting methods, extend `Downcast` or `DowncastSync`
+//! // and run `impl_downcast!()` on the trait.
+//! trait Base: DowncastSync {}
+//! impl_downcast!(sync Base);  // `sync` => also produce `Arc` downcasts.
 //!
 //! // Concrete types implementing Base.
 //! #[derive(Debug)]
@@ -81,6 +92,14 @@
 //!     let base = res.unwrap_err();
 //!     // Convert `Box<Base>` into `Box<Foo>`.
 //!     assert_eq!(42, base.downcast::<Foo>().map_err(|_| "Shouldn't happen.").unwrap().0);
+//!
+//!     // Also works with `Rc`.
+//!     let mut rc: Rc<Base> = Rc::new(Foo(42));
+//!     assert_eq!(42, rc.downcast_rc::<Foo>().map_err(|_| "Shouldn't happen.").unwrap().0);
+//!
+//!     // Since this trait is `Sync`, it also supports `Arc` downcasts.
+//!     let mut arc: Arc<Base> = Arc::new(Foo(42));
+//!     assert_eq!(42, arc.downcast_arc::<Foo>().map_err(|_| "Shouldn't happen.").unwrap().0);
 //! }
 //! ```
 //!
@@ -91,8 +110,8 @@
 //! extern crate downcast_rs;
 //! use downcast_rs::Downcast;
 //!
-//! // To create a trait with downcasting methods, extend `Downcast` and run
-//! // `impl_downcast!()` on the trait.
+//! // To create a trait with downcasting methods, extend `Downcast` or `DowncastSync`
+//! // and run `impl_downcast!()` on the trait.
 //! trait Base<T: Clone>: Downcast { type H: Copy; }
 //! downcast_rs::impl_downcast!(Base<T> assoc H where T: Clone, H: Copy);
 //! // or: impl_downcast!(concrete Base<u32> assoc H=f32)
@@ -125,6 +144,9 @@ pub trait Downcast: Any {
     /// Convert `Box<Trait>` (where `Trait: Downcast`) to `Box<Any>`. `Box<Any>` can then be
     /// further `downcast` into `Box<ConcreteType>` where `ConcreteType` implements `Trait`.
     fn into_any(self: Box<Self>) -> Box<Any>;
+    /// Convert `Rc<Trait>` (where `Trait: Downcast`) to `Rc<Any>`. `Rc<Any>` can then be
+    /// further `downcast` into `Rc<ConcreteType>` where `ConcreteType` implements `Trait`.
+    fn into_any_rc(self: ::std::rc::Rc<Self>) -> ::std::rc::Rc<Any>;
     /// Convert `&Trait` (where `Trait: Downcast`) to `&Any`. This is needed since Rust cannot
     /// generate `&Any`'s vtable from `&Trait`'s.
     fn as_any(&self) -> &Any;
@@ -135,8 +157,20 @@ pub trait Downcast: Any {
 
 impl<T: Any> Downcast for T {
     fn into_any(self: Box<Self>) -> Box<Any> { self }
+    fn into_any_rc(self: ::std::rc::Rc<Self>) -> ::std::rc::Rc<Any> { self }
     fn as_any(&self) -> &Any { self }
     fn as_any_mut(&mut self) -> &mut Any { self }
+}
+
+/// Extends `Downcast` to support `Sync` traits that thus support `Arc` downcasting as well.
+pub trait DowncastSync: Downcast + Send + Sync {
+    /// Convert `Arc<Trait>` (where `Trait: Downcast`) to `Arc<Any>`. `Arc<Any>` can then be
+    /// further `downcast` into `Arc<ConcreteType>` where `ConcreteType` implements `Trait`.
+    fn into_any_arc(self: ::std::sync::Arc<Self>) -> ::std::sync::Arc<Any + Send + Sync>;
+}
+
+impl<T: Any + Send + Sync> DowncastSync for T {
+    fn into_any_arc(self: ::std::sync::Arc<Self>) -> ::std::sync::Arc<Any + Send + Sync> { self }
 }
 
 /// Adds downcasting support to traits that extend `downcast::Downcast` by defining forwarding
@@ -162,6 +196,23 @@ macro_rules! impl_downcast {
         }
     };
 
+    (@impl_full_sync
+        $trait_:ident [$($param_types:tt)*]
+        for [$($forall_types:ident),*]
+        where [$($preds:tt)*]
+    ) => {
+        impl_downcast! {
+            @inject_where
+                [impl<$($forall_types),*> $trait_<$($param_types)*>]
+                types [$($forall_types),*]
+                where [$($preds)*]
+                [{
+                    impl_downcast! { @impl_body $trait_ [$($param_types)*] }
+                    impl_downcast! { @impl_body_sync $trait_ [$($param_types)*] }
+                }]
+        }
+    };
+
     (@impl_body $trait_:ident [$($types:tt)*]) => {
         /// Returns true if the trait object wraps an object of type `__T`.
         #[inline]
@@ -180,6 +231,18 @@ macro_rules! impl_downcast {
                 Err(self)
             }
         }
+        /// Returns an `Rc`-ed object from an `Rc`-ed trait object if the underlying object is of
+        /// type `__T`. Returns the original `Rc`-ed trait if it isn't.
+        #[inline]
+        pub fn downcast_rc<__T: $trait_<$($types)*>>(
+            self: ::std::rc::Rc<Self>
+        ) -> ::std::result::Result<::std::rc::Rc<__T>, ::std::rc::Rc<Self>> {
+            if self.is::<__T>() {
+                Ok($crate::Downcast::into_any_rc(self).downcast::<__T>().unwrap())
+            } else {
+                Err(self)
+            }
+        }
         /// Returns a reference to the object within the trait object if it is of type `__T`, or
         /// `None` if it isn't.
         #[inline]
@@ -191,6 +254,23 @@ macro_rules! impl_downcast {
         #[inline]
         pub fn downcast_mut<__T: $trait_<$($types)*>>(&mut self) -> ::std::option::Option<&mut __T> {
             $crate::Downcast::as_any_mut(self).downcast_mut::<__T>()
+        }
+    };
+
+    (@impl_body_sync $trait_:ident [$($types:tt)*]) => {
+        /// Returns an `Arc`-ed object from an `Arc`-ed trait object if the underlying object is of
+        /// type `__T`. Returns the original `Arc`-ed trait if it isn't.
+        #[inline]
+        pub fn downcast_arc<__T: $trait_<$($types)*>>(
+            self: ::std::sync::Arc<Self>,
+        ) -> ::std::result::Result<::std::sync::Arc<__T>, ::std::sync::Arc<Self>>
+            where __T: ::std::any::Any + ::std::marker::Send + ::std::marker::Sync
+        {
+            if self.is::<__T>() {
+                Ok($crate::DowncastSync::into_any_arc(self).downcast::<__T>().unwrap())
+            } else {
+                Err(self)
+            }
         }
     };
 
@@ -222,26 +302,48 @@ macro_rules! impl_downcast {
     // No type parameters.
     ($trait_:ident   ) => { impl_downcast! { @impl_full $trait_ [] for [] where [] } };
     ($trait_:ident <>) => { impl_downcast! { @impl_full $trait_ [] for [] where [] } };
+    (sync $trait_:ident   ) => { impl_downcast! { @impl_full_sync $trait_ [] for [] where [] } };
+    (sync $trait_:ident <>) => { impl_downcast! { @impl_full_sync $trait_ [] for [] where [] } };
     // Type parameters.
     ($trait_:ident < $($types:ident),* >) => {
         impl_downcast! { @impl_full $trait_ [$($types),*] for [$($types),*] where [] }
+    };
+    (sync $trait_:ident < $($types:ident),* >) => {
+        impl_downcast! { @impl_full_sync $trait_ [$($types),*] for [$($types),*] where [] }
     };
     // Type parameters and where clauses.
     ($trait_:ident < $($types:ident),* > where $($preds:tt)+) => {
         impl_downcast! { @impl_full $trait_ [$($types),*] for [$($types),*] where [$($preds)*] }
     };
+    (sync $trait_:ident < $($types:ident),* > where $($preds:tt)+) => {
+        impl_downcast! { @impl_full_sync $trait_ [$($types),*] for [$($types),*] where [$($preds)*] }
+    };
     // Associated types.
     ($trait_:ident assoc $($atypes:ident),*) => {
         impl_downcast! { @impl_full $trait_ [$($atypes = $atypes),*] for [$($atypes),*] where [] }
+    };
+    (sync $trait_:ident assoc $($atypes:ident),*) => {
+        impl_downcast! { @impl_full_sync $trait_ [$($atypes = $atypes),*] for [$($atypes),*] where [] }
     };
     // Associated types and where clauses.
     ($trait_:ident assoc $($atypes:ident),* where $($preds:tt)+) => {
         impl_downcast! { @impl_full $trait_ [$($atypes = $atypes),*] for [$($atypes),*] where [$($preds)*] }
     };
+    (sync $trait_:ident assoc $($atypes:ident),* where $($preds:tt)+) => {
+        impl_downcast! { @impl_full_sync $trait_ [$($atypes = $atypes),*] for [$($atypes),*] where [$($preds)*] }
+    };
     // Type parameters and associated types.
     ($trait_:ident < $($types:ident),* > assoc $($atypes:ident),*) => {
         impl_downcast! {
             @impl_full
+                $trait_ [$($types),*, $($atypes = $atypes),*]
+                for [$($types),*, $($atypes),*]
+                where []
+        }
+    };
+    (sync $trait_:ident < $($types:ident),* > assoc $($atypes:ident),*) => {
+        impl_downcast! {
+            @impl_full_sync
                 $trait_ [$($types),*, $($atypes = $atypes),*]
                 for [$($types),*, $($atypes),*]
                 where []
@@ -256,17 +358,34 @@ macro_rules! impl_downcast {
                 where [$($preds)*]
         }
     };
+    (sync $trait_:ident < $($types:ident),* > assoc $($atypes:ident),* where $($preds:tt)+) => {
+        impl_downcast! {
+            @impl_full_sync
+                $trait_ [$($types),*, $($atypes = $atypes),*]
+                for [$($types),*, $($atypes),*]
+                where [$($preds)*]
+        }
+    };
     // Concretely-parametrized types.
     (concrete $trait_:ident < $($types:ident),* >) => {
         impl_downcast! { @impl_full $trait_ [$($types),*] for [] where [] }
+    };
+    (sync concrete $trait_:ident < $($types:ident),* >) => {
+        impl_downcast! { @impl_full_sync $trait_ [$($types),*] for [] where [] }
     };
     // Concretely-associated types types.
     (concrete $trait_:ident assoc $($atypes:ident = $aty:ty),*) => {
         impl_downcast! { @impl_full $trait_ [$($atypes = $aty),*] for [] where [] }
     };
+    (sync concrete $trait_:ident assoc $($atypes:ident = $aty:ty),*) => {
+        impl_downcast! { @impl_full_sync $trait_ [$($atypes = $aty),*] for [] where [] }
+    };
     // Concretely-parametrized types with concrete associated types.
     (concrete $trait_:ident < $($types:ident),* > assoc $($atypes:ident = $aty:ty),*) => {
         impl_downcast! { @impl_full $trait_ [$($types),*, $($atypes = $aty),*] for [] where [] }
+    };
+    (sync concrete $trait_:ident < $($types:ident),* > assoc $($atypes:ident = $aty:ty),*) => {
+        impl_downcast! { @impl_full_sync $trait_ [$($types),*, $($atypes = $aty),*] for [] where [] }
     };
 }
 
@@ -275,13 +394,81 @@ macro_rules! impl_downcast {
 mod test {
     macro_rules! test_mod {
         (
-            $test_name:ident,
+            $test_mod_name:ident,
+            trait $base_trait:path { $($base_impl:tt)* },
+            non_sync: { $($non_sync_def:tt)+ },
+            sync: { $($sync_def:tt)+ }
+        ) => {
+            test_mod! {
+                $test_mod_name,
+                trait $base_trait { $($base_impl:tt)* },
+                type $base_trait,
+                non_sync: { $($non_sync_def)* },
+                sync: { $($sync_def)* }
+            }
+        };
+
+        (
+            $test_mod_name:ident,
             trait $base_trait:path { $($base_impl:tt)* },
             type $base_type:ty,
-            { $($def:tt)+ }
+            non_sync: { $($non_sync_def:tt)+ },
+            sync: { $($sync_def:tt)+ }
         ) => {
-            mod $test_name {
-                use super::super::Downcast;
+            mod $test_mod_name {
+                test_mod!(
+                    @test
+                    $test_mod_name,
+                    test_name: test_non_sync,
+                    trait $base_trait { $($base_impl)* },
+                    type $base_type,
+                    { $($non_sync_def)+ },
+                    []);
+
+                test_mod!(
+                    @test
+                    $test_mod_name,
+                    test_name: test_sync,
+                    trait $base_trait { $($base_impl)* },
+                    type $base_type,
+                    { $($sync_def)+ },
+                    [{
+                        // Fail to convert Arc<Base> into Arc<Bar>.
+                        let arc: ::std::sync::Arc<$base_type> = ::std::sync::Arc::new(Foo(42));
+                        let res = arc.downcast_arc::<Bar>();
+                        assert!(res.is_err());
+                        let arc = res.unwrap_err();
+                        // Convert Arc<Base> into Arc<Foo>.
+                        assert_eq!(
+                            42, arc.downcast_arc::<Foo>().map_err(|_| "Shouldn't happen.").unwrap().0);
+                    }]);
+            }
+        };
+
+        (
+            @test
+            $test_mod_name:ident,
+            test_name: $test_name:ident,
+            trait $base_trait:path { $($base_impl:tt)* },
+            type $base_type:ty,
+            { $($def:tt)+ },
+            [ $($more_tests:block)* ]
+        ) => {
+            #[test]
+            fn $test_name() {
+                #[allow(unused_imports)]
+                use super::super::{Downcast, DowncastSync};
+
+                // Should work even if standard objects (especially those in the prelude) are
+                // aliased to something else.
+                #[allow(dead_code)] struct Any;
+                #[allow(dead_code)] struct Arc;
+                #[allow(dead_code)] struct Box;
+                #[allow(dead_code)] struct Option;
+                #[allow(dead_code)] struct Result;
+                #[allow(dead_code)] struct Rc;
+                #[allow(dead_code)] struct Send;
+                #[allow(dead_code)] struct Sync;
 
                 // A trait that can be downcast.
                 $($def)*
@@ -307,96 +494,140 @@ mod test {
                     }
                 }
 
-                #[test]
-                fn test() {
-                    let mut base: ::std::boxed::Box<$base_type> = ::std::boxed::Box::new(Foo(42));
-                    assert_eq!(get_val(&base), 42);
+                let mut base: ::std::boxed::Box<$base_type> = ::std::boxed::Box::new(Foo(42));
+                assert_eq!(get_val(&base), 42);
 
-                    // Try sequential downcasts.
-                    if let Some(foo) = base.downcast_ref::<Foo>() {
-                        assert_eq!(foo.0, 42);
-                    } else if let Some(bar) = base.downcast_ref::<Bar>() {
-                        assert_eq!(bar.0, 42.0);
-                    }
-
-                    set_val(&mut base, 6*9);
-                    assert_eq!(get_val(&base), 6*9);
-
-                    assert!(base.is::<Foo>());
-
-                    // Fail to convert Box<Base> into Box<Bar>.
-                    let res = base.downcast::<Bar>();
-                    assert!(res.is_err());
-                    let base = res.unwrap_err();
-                    // Convert Box<Base> into Box<Foo>.
-                    assert_eq!(
-                        6*9, base.downcast::<Foo>().map_err(|_| "Shouldn't happen.").unwrap().0);
+                // Try sequential downcasts.
+                if let Some(foo) = base.downcast_ref::<Foo>() {
+                    assert_eq!(foo.0, 42);
+                } else if let Some(bar) = base.downcast_ref::<Bar>() {
+                    assert_eq!(bar.0, 42.0);
                 }
+
+                set_val(&mut base, 6*9);
+                assert_eq!(get_val(&base), 6*9);
+
+                assert!(base.is::<Foo>());
+
+                // Fail to convert Box<Base> into Box<Bar>.
+                let res = base.downcast::<Bar>();
+                assert!(res.is_err());
+                let base = res.unwrap_err();
+                // Convert Box<Base> into Box<Foo>.
+                assert_eq!(
+                    6*9, base.downcast::<Foo>().map_err(|_| "Shouldn't happen.").unwrap().0);
+
+                // Fail to convert Rc<Base> into Rc<Bar>.
+                let rc: ::std::rc::Rc<$base_type> = ::std::rc::Rc::new(Foo(42));
+                let res = rc.downcast_rc::<Bar>();
+                assert!(res.is_err());
+                let rc = res.unwrap_err();
+                // Convert Rc<Base> into Rc<Foo>.
+                assert_eq!(
+                    42, rc.downcast_rc::<Foo>().map_err(|_| "Shouldn't happen.").unwrap().0);
+
+                $($more_tests)*
             }
         };
-
-        (
-            $test_name:ident,
-            trait $base_trait:path { $($base_impl:tt)* },
-            { $($def:tt)+ }
-        ) => {
-            test_mod! {
-                $test_name, trait $base_trait { $($base_impl:tt)* }, type $base_trait, { $($def)* }
-            }
-        }
     }
 
-    test_mod!(non_generic, trait Base {}, {
-        trait Base: Downcast {}
-        impl_downcast!(Base);
-    });
+    test_mod!(non_generic, trait Base {},
+        non_sync: {
+            trait Base: Downcast {}
+            impl_downcast!(Base);
+        },
+        sync: {
+            trait Base: DowncastSync {}
+            impl_downcast!(sync Base);
+        });
 
-    test_mod!(generic, trait Base<u32> {}, {
-        trait Base<T>: Downcast {}
-        impl_downcast!(Base<T>);
-    });
+    test_mod!(generic, trait Base<u32> {},
+        non_sync: {
+            trait Base<T>: Downcast {}
+            impl_downcast!(Base<T>);
+        },
+        sync: {
+            trait Base<T>: DowncastSync {}
+            impl_downcast!(sync Base<T>);
+        });
 
-    test_mod!(constrained_generic, trait Base<u32> {}, {
-        // Should work even if standard objects in the prelude are aliased to something else.
-        #[allow(dead_code)] struct Box;
-        #[allow(dead_code)] struct Option;
-        #[allow(dead_code)] struct Result;
-        trait Base<T: Copy>: Downcast {}
-        impl_downcast!(Base<T> where T: Copy);
-    });
+    test_mod!(constrained_generic, trait Base<u32> {},
+        non_sync: {
+            trait Base<T: Copy>: Downcast {}
+            impl_downcast!(Base<T> where T: Copy);
+        },
+        sync: {
+            trait Base<T: Copy>: DowncastSync {}
+            impl_downcast!(sync Base<T> where T: Copy);
+        });
 
-    test_mod!(associated, trait Base { type H = f32; }, type Base<H=f32>, {
-        trait Base: Downcast { type H; }
-        impl_downcast!(Base assoc H);
-    });
+    test_mod!(associated, trait Base { type H = f32; }, type Base<H=f32>,
+        non_sync: {
+            trait Base: Downcast { type H; }
+            impl_downcast!(Base assoc H);
+        },
+        sync: {
+            trait Base: DowncastSync { type H; }
+            impl_downcast!(sync Base assoc H);
+        });
 
-    test_mod!(constrained_associated, trait Base { type H = f32; }, type Base<H=f32>, {
-        trait Base: Downcast { type H: Copy; }
-        impl_downcast!(Base assoc H where H: Copy);
-    });
+    test_mod!(constrained_associated, trait Base { type H = f32; }, type Base<H=f32>,
+        non_sync: {
+            trait Base: Downcast { type H: Copy; }
+            impl_downcast!(Base assoc H where H: Copy);
+        },
+        sync: {
+            trait Base: DowncastSync { type H: Copy; }
+            impl_downcast!(sync Base assoc H where H: Copy);
+        });
 
-    test_mod!(param_and_associated, trait Base<u32> { type H = f32; }, type Base<u32, H=f32>, {
-        trait Base<T>: Downcast { type H; }
-        impl_downcast!(Base<T> assoc H);
-    });
+    test_mod!(param_and_associated, trait Base<u32> { type H = f32; }, type Base<u32, H=f32>,
+        non_sync: {
+            trait Base<T>: Downcast { type H; }
+            impl_downcast!(Base<T> assoc H);
+        },
+        sync: {
+            trait Base<T>: DowncastSync { type H; }
+            impl_downcast!(sync Base<T> assoc H);
+        });
 
-    test_mod!(constrained_param_and_associated, trait Base<u32> { type H = f32; }, type Base<u32, H=f32>, {
-        trait Base<T: Clone>: Downcast { type H: Copy; }
-        impl_downcast!(Base<T> assoc H where T: Clone, H: Copy);
-    });
+    test_mod!(constrained_param_and_associated, trait Base<u32> { type H = f32; }, type Base<u32, H=f32>,
+        non_sync: {
+            trait Base<T: Clone>: Downcast { type H: Copy; }
+            impl_downcast!(Base<T> assoc H where T: Clone, H: Copy);
+        },
+        sync: {
+            trait Base<T: Clone>: DowncastSync { type H: Copy; }
+            impl_downcast!(sync Base<T> assoc H where T: Clone, H: Copy);
+        });
 
-    test_mod!(concrete_parametrized, trait Base<u32> {}, {
-        trait Base<T>: Downcast {}
-        impl_downcast!(concrete Base<u32>);
-    });
+    test_mod!(concrete_parametrized, trait Base<u32> {},
+        non_sync: {
+            trait Base<T>: Downcast {}
+            impl_downcast!(concrete Base<u32>);
+        },
+        sync: {
+            trait Base<T>: DowncastSync {}
+            impl_downcast!(sync concrete Base<u32>);
+        });
 
-    test_mod!(concrete_associated, trait Base { type H = u32; }, type Base<H=u32>, {
-        trait Base: Downcast { type H; }
-        impl_downcast!(concrete Base assoc H=u32);
-    });
+    test_mod!(concrete_associated, trait Base { type H = u32; }, type Base<H=u32>,
+        non_sync: {
+            trait Base: Downcast { type H; }
+            impl_downcast!(concrete Base assoc H=u32);
+        },
+        sync: {
+            trait Base: DowncastSync { type H; }
+            impl_downcast!(sync concrete Base assoc H=u32);
+        });
 
-    test_mod!(concrete_parametrized_associated, trait Base<u32> { type H = f32; }, type Base<u32, H=f32>, {
-        trait Base<T>: Downcast { type H; }
-        impl_downcast!(concrete Base<u32> assoc H=f32);
-    });
+    test_mod!(concrete_parametrized_associated, trait Base<u32> { type H = f32; }, type Base<u32, H=f32>,
+        non_sync: {
+            trait Base<T>: Downcast { type H; }
+            impl_downcast!(concrete Base<u32> assoc H=f32);
+        },
+        sync: {
+            trait Base<T>: DowncastSync { type H; }
+            impl_downcast!(sync concrete Base<u32> assoc H=f32);
+        });
 }